move stuff inside web app folder
