+++ web_app/public/manifest.json
