+++ web_app/src/Button.js
@@ -1,0 +1,20 @@
+import styled from 'styled-components';
+import style from 'styled-theming';
+
+const getBackground = style.variants('mode', 'variant', {
+  normal: {
+    light: '#EEE',
+    dark: '#EEE'
+  },
+  primary: {
+    light: 'papayawhip',
+    dark: 'pink'
+  }
+});
+
+const Button = styled.button`
+  background-color: ${getBackground};
+  margin-left : 1300px;
+`;
+
+export default Button;
