+++ web_app/src/sidebar.css
